requirements.txt
Flask
Flask-Talisman
Flask-CORS
numpy
Pillow
Werkzeug
gunicorn
nose2[coverage-plugin]
coverage
flake8
pylint

